Now G function is generalized and works fine. Look at GT GT function for modification now.

+++ module_data_type.f90
@@ -319,7 +319,7 @@ IMPLICIT NONE
         INTEGER,DIMENSION(:),ALLOCATABLE        :: i_udof_p
         INTEGER,DIMENSION(:,:),ALLOCATABLE      :: kinmap
         REAL(dp),DIMENSION(:,:),ALLOCATABLE     :: kindata
-        INTEGER,DIMENSION(:,:),ALLOCATABLE      :: P_map
+        REAL(dp),DIMENSION(:,:),ALLOCATABLE     :: P_map
     END TYPE
